+++ lambdas/src/worker/package.json
@@ -1,4 +1,10 @@
 {
+  "name": "worker",
+  "version": "1.0.0",
+  "main": "src/index.ts",
+  "scripts": {
+    "build": "npx webpack && (cp ./package.json dist/package.json && cd dist && npm i --production)"
+  },
   "dependencies": {
     "chrome-aws-lambda": "^2.0.2",
     "chrome-launcher": "^0.12.0",
@@ -7,5 +13,19 @@
     "lighthouse": "^5.6.0",
     "lighthouse-logger": "^1.2.0",
     "puppeteer-core": "^2.1.1"
+  },
+  "devDependencies": {
+    "@types/aws-sdk": "^2.7.0",
+    "@types/node": "^13.7.1",
+    "@types/puppeteer": "^2.0.0",
+    "@types/puppeteer-core": "^2.0.0",
+    "awesome-typescript-loader": "^5.2.1",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "terser-webpack-plugin": "^2.3.4",
+    "typescript": "^3.7.5",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11",
+    "webpack-node-externals": "^1.7.2"
   }
 }
