+++ lambdas/src/worker/src/@types/lighthouse/index.d.ts
@@ -1,0 +1,2 @@
+// I needed this as a shim until Lighthouse gets their bindings in order
+declare module 'lighthouse';
